Set s3 url when we create catalog server (#11)
Previously, for the convenience of testing, the system used the local s3 service by default, 
which obviously cannot meet the requirements of distributed deployment. Now we set the
url of the s3 service when initializing the catalog server, and this url will be distributed to 
the corresponding computing process. In this way, we can set a different s3 service for 
each catalog server and make it accessible to the entire virtual machine group.